+++ src/components/Navbars/RTLNavbar.js
@@ -98,7 +98,7 @@ function AdminNavbar(props) {
                   nav
                 >
                   <div className="notification d-none d-lg-block d-xl-block" />
-                  <i className="tim-icons icon-sound-wave" />
+                  {/* <i className="tim-icons icon-sound-wave" /> */}
                   <p className="d-lg-none">Notifications</p>
                 </DropdownToggle>
                 <DropdownMenu className="dropdown-navbar" tag="ul" right>